requirements.txt
fastapi==0.99.0
Pillow==9.5.0
pydantic==1.10.9
Requests==2.31.0
uvicorn==0.22.0

